Move domain::OptionTags::DomainCreator to DomainCreator library
