+++ src/Domain/CMakeLists.txt
@@ -47,7 +47,6 @@ spectre_target_headers(
   InterfaceLogicalCoordinates.hpp
   JacobianDiagnostic.hpp
   MinimumGridSpacing.hpp
-  OptionTags.hpp
   SizeOfElement.hpp
   SurfaceJacobian.hpp
   Tags.hpp
